+++ composeApp/src/commonMain/kotlin/io/github/sophon/fightingnerd/AppConstants.kt
@@ -1,3 +1,6 @@
 package io.github.sophon.fightingnerd
 
 internal const val KEY_FIRST_TIME_HOME_INIT_DONE = "first_time_home_init_done"
+
+internal const val INTERVAL_UPDATE_DEFAULT_D = 7
+