+++ composeApp/src/commonMain/kotlin/io/github/sophon/fightingnerd/core/data/PreferenceRepo.kt
@@ -4,8 +4,12 @@ import io.github.sophon.core.architecture.EmptyResult
 import io.github.sophon.fightingnerd.core.model.AppError
 import io.github.sophon.fightingnerd.theme.ThemeMode
 import kotlinx.coroutines.flow.Flow
+import kotlin.time.Duration
 
 internal interface PreferenceRepo {
     fun subscribeToTheme(): Flow<ThemeMode>
     suspend fun setTheme(themeMode: ThemeMode): EmptyResult<AppError>
+
+    fun subscribeToUpdateInterval(): Flow<Duration?>
+    suspend fun setUpdateInterval(duration: Duration?): EmptyResult<AppError>
 }